Introduce federating and federated catalog
Preliminary text referring the "federated" and "federating" catalogues and the possible support for two-step search by relaxing the href URL constraint from OAFeat for the /items endpoints to allow the same endpoint to be used by both federating and federated catalog.   Both "federating" and "federated" catalogues may eventually be subject to different (minimal) requirements for a federated scenario to work.